requirements.txt
bidict==0.21.4
biosppy==0.8.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.30.0
h5py==3.6.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
shortuuid==1.0.8
six==1.16.0
threadpoolctl==3.1.0
torch==1.11.0
typing-extensions==4.1.1
urllib3==1.26.8
wfdb==3.4.1
ipython==8.3.0
tqdm==4.64.0
ipykernel==6.13.0
typeguard==2.13.3
mlflow==1.25.1
tensorflow==2.8.0
torchvision==0.12.0
pexpect==4.8.0
paramiko==2.10.4 
pysftp==0.2.9 
fastai==2.6.3 